Moved donate command entirely to profile.js.
